Add option to run per-detector code in separate processes
