+++ python/lsst/l1dbproto/__init__.py
@@ -1,4 +1,3 @@
 
 from .version import *
 from .config import *
-from .ap_proto import *
